Change usage of labels.

+++ src/cpu/process.cpp
@@ -12,17 +12,14 @@ static void cpu_dump(int *cmd_buf, size_t cmd_count, size_t ip)
 
         for (size_t i = 0; i < cmd_count; i++)
                 printf("%02zu ", i);
-
         printf("\n");
 
         for (size_t i = 0; i < cmd_count; i++)
                 printf("%02d ", cmd_buf[i]);
-
         printf("\n");
 
         for (size_t i = 0; i < ip; i++)
                 printf("   ");
-
         printf("^\n");
 }
 
@@ -64,7 +61,7 @@ execute(int *cmd_buf, size_t size)
                         DEF_CMD(DIV, POP(val1) POP(val2) PUSH(val2 / val1))
                         DEF_CMD(DUP, POP(val1) PUSH(val1) PUSH(val2))
                         DEF_CMD(OUT, POP(val1) printf("%d\n", val1);)
-                        DEF_CMD(JMP, ip = cmd_buf[ip];)
+                        DEF_CMD(JMP, ip = (size_t) cmd_buf[ip];)
                         DEF_CMD(IN, scanf("%d", &val1); PUSH(val1))
                         DEF_CMD(SQRT, val1 = (int) sqrt(val1); PUSH(val1))
                         DEF_CMD(DMP & CMD_MASK, cpu_dump(cmd_buf, size, ip);)
